+++ docs/CHANGELOG.md
@@ -17,6 +17,8 @@
 - `[Column Chart]` Added support for double click to Column, Column Grouped, Column Stacked, Column Stacked-singular and Column Positive Negative. ([#3229](https://github.com/infor-design/enterprise/issues/3229))
 - `[Datagrid]` Added api setting `allowChildExpandOnMatchOnly` with Datagrid. It will show/hide children match only or all of them this setting only will effect if use with `allowChildExpandOnMatch:true`. ([#4209](https://github.com/infor-design/enterprise/issues/4209))
 - `[Datagrid]` Fixed a bug where filter dropdown menus did not close when focusing a filter input. ([#4766](https://github.com/infor-design/enterprise/issues/4766))
+- `[Datagrid]` Fixed an issue where the keyboard was not working to sort data for sortable columns. ([#4858](https://github.com/infor-design/enterprise/issues/4858))
+- `[Datagrid]` Fixed an issue where the keyboard was not working to select all from header checkbox. ([#4859](https://github.com/infor-design/enterprise/issues/4859))
 - `[Datagrid]` Fixed an issue where the selection was getting clear after use pagesize dropdown for client side paging. ([#4915](https://github.com/infor-design/enterprise/issues/4915))
 - `[Datagrid]` Fixed an error seen clicking items if using a flex toolbar for the datagrid toolbar. ([#4941](https://github.com/infor-design/enterprise/issues/4941))
 - `[Datagrid]` Only show row status when dirty indicator and row status both exist to address conflicting visual issue. ([#4918](https://github.com/infor-design/enterprise/issues/4918))